🐛 fix: separate authOptions from route handler
